requirements.txt
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
openai>=1.1.1
httpx>=0.25.1
python-dotenv>=1.0.0
pydantic>=2.5.0
starlette>=0.35.0
pytest>=7.4.0
black>=23.9.1
ruff>=0.0.292
pytest-asyncio>=0.21.1
